+++ src/app.module.ts
@@ -1,19 +1,29 @@
 import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
 import { createObserveModule } from '@nestjs/observe';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import configuration from './config/configuration';
+import { DatabaseModule } from './database/database.module';
+import { UsersModule } from './users/users.module';
 
-export const { ObserveModule, ObserveInstrument } = createObserveModule();
+// export const { ObserveModule, ObserveInstrument } = createObserveModule();
 
 @Module({
   imports: [
-    // Distributed tracing, auto-correlated logs, request/job metrics, error
-    // telemetry, alarms, and more — out of the box. Sign up at https://observe.nestjs.com
-    ObserveModule.forRoot({
-      appKey: 'YOUR_APP_KEY',
-      appSecret: 'YOUR_APP_SECRET',
-      serviceId: 'wallet-back',
+    ConfigModule.forRoot({
+      isGlobal: true,
+      load: [configuration],
     }),
+    // Distributed tracing, auto-correlated logs, request/job metrics, error
+    // telemetry, alarms, e mais — pronto para usar. Cadastre-se em https://observe.nestjs.com
+    // ObserveModule.forRoot({
+    //   appKey: 'YOUR_APP_KEY',
+    //   appSecret: 'YOUR_APP_SECRET',
+    //   serviceId: 'wallet-back',
+    // }),
+    DatabaseModule,
+    UsersModule,
   ],
   controllers: [AppController],
   providers: [AppService],