+++ src/common/enums/user-status.enum.ts
@@ -1,5 +1,4 @@
 export enum UserStatus {
-  PENDING_VERIFICATION = 'PENDING_VERIFICATION',
   ACTIVE = 'ACTIVE',
   BLOCKED = 'BLOCKED',
 }